Switch to built-in kubernetes customization from our custom templates
I don't know if anyone else is still using this repository at this point,
but generating kubernetes configs from templates every time I change something showed to be annoying and error-prone.

I will make a branch `templates` that points to the state prior to this commit, so you can use that if you prefer the old behaviour.